+++ ft_memcmp.c
@@ -9,9 +9,9 @@ int ft_memcmp(const void *s1, const void *s2, size_t n)
     str1 = (char *)s1;
     str2 = (char *)s2;
     i = 0;
-    if (n == 0)
+    if ((int)n == 0)
         return (0);
-    while (str1[i] == str2[i] && i < n)
+    while (str1[i] == str2[i] && i < (int)n)
     {
         i++;
     }
