+++ ft_memcpy.c
@@ -9,7 +9,7 @@ void *ft_memcpy(void *dest, const void *src, size_t n)
     i = 0;
     d = (unsigned char *)dest;
     s = (unsigned char *)src;
-    while (i < n)
+    while (i < (int)n)
     {
         d[i] = s[i];
         i++;
