+++ ft_strncmp.c
@@ -7,9 +7,9 @@ int ft_strncmp(const char *s1, const char *s2, size_t n)
     i = 0;
     if (n == 0)
         return (0);
-    while (s1[i] && s2[i] && s1[i] == s2[i] && i < n)
+    while (s1[i] && s2[i] && s1[i] == s2[i] && i < (int)n)
         i++;
-    if (i == n)
+    if (i == (int)n)
         i--;
     return (s1[i] - s2[i]);
 }
